% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_markdown.r
\name{read_markdown}
\alias{read_markdown}
\title{Reads Markdown tables}
\usage{
read_markdown(markdown, lower_header = TRUE)
}
\arguments{
\item{markdown}{a Markdown table}

\item{lower_header}{an optional argument controlling whether or not the header should be lowercase}
}
\value{
a data.frame
}
\description{
Reads Markdown tables
}
\examples{
# define table
markdown_table<-"
| state | capital | area |
| --- | --- | --- |
| California | Sacramento | 163700 |
| Massachusetts | Boston | 10565 |
| Illinois | Springfield | 57914 |
"

# read table
read_markdown(markdown_table)
}
