% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_markdown.r
\name{read_markdown}
\alias{read_markdown}
\title{\code{read_markdown()}}
\usage{
read_markdown(markdown, lower_header = TRUE)
}
\arguments{
\item{markdown}{a Markdown table}

\item{lower_header}{an optional argument controlling whether or not the header should be lowercase}
}
\value{
a data.frame
}
\description{
\code{read_markdown()}
}
\examples{
markdown<-"
fruit| price
-----|-----
apple|2.05
pear|1.37
orange|3.09
"
read_markdown(markdown)
}
